Make suggested changes, build currently failing
