+++ extensions/ql-vscode/test/sarif/emptyResultsSarif.sarif
