+++ extensions/ql-vscode/test/sarif/invalidSarif.sarif
