+++ extensions/ql-vscode/test/sarif/validSarif.sarif
